Free play.

+++ OpenGL.Hal/BufferObject.cs
@@ -17,6 +17,7 @@
 // USA
 
 using System;
+using System.Diagnostics;
 using System.Runtime.InteropServices;
 
 namespace OpenGL
@@ -29,7 +30,7 @@ namespace OpenGL
 	/// Any data that is transferred from and to the GPU shall be organized into a BufferObject instances.
 	/// </para>
 	/// <para>
-	/// This clas actually stores the data using a <see cref="AlignedMemoryBuffer"/>, indeed un unmanaged
+	/// This class actually stores the data using a <see cref="AlignedMemoryBuffer"/>, indeed un unmanaged
 	/// chunk of memory (aligned to 16 byte boundaries).
 	/// </para>
 	/// <para>
@@ -48,9 +49,9 @@ namespace OpenGL
 		/// A <see cref="BufferTargetARB"/> that specify the buffer object type.
 		/// </param>
 		/// <param name="hint">
-		/// An <see cref="Hint"/> that specify the data buffer usage hints.
+		/// An <see cref="BufferObjectHint"/> that specify the data buffer usage hints.
 		/// </param>
-		protected BufferObject(BufferTargetARB type, Hint hint)
+		protected BufferObject(BufferTargetARB type, BufferObjectHint hint)
 		{
 			// Store the buffer object type
 			_Type = type;
@@ -60,83 +61,7 @@ namespace OpenGL
 		
 		#endregion
 
-		#region Buffer Data Definition
-
-		/// <summary>
-		/// Buffer usage hints.
-		/// </summary>
-		/// <remarks>
-		/// <para>
-		/// This enumeration determine the BufferObject usage. It determines whether client memory shall be allocated
-		/// and released.
-		/// </para>
-		/// <para>
-		/// Additionally give an hint about the buffer basic operation (read, copy or draw) and ratio between source
-		/// updates and draw updates.
-		/// </para>
-		/// </remarks>
-		public enum Hint
-		{
-			/// <summary>
-			/// Buffer used for drawing, specified once by application data.
-			/// </summary>
-			/// <remarks>
-			/// <para>
-			/// By using this hint, this BufferObject allocates client memory to set buffer object contents,
-			/// and after done it release the client memory.
-			/// </para>
-			/// </remarks>
-			StaticCpuDraw = Gl.STATIC_DRAW,
-			/// <summary>
-			/// Buffer used for drawing, specified once by OpenGL rendering.
-			/// </summary>
-			/// <remarks>
-			/// <para>
-			/// By using this hint, this BufferObject does not allocates client memory.
-			/// </para>
-			/// </remarks>
-			StaticGpuDraw = Gl.STATIC_COPY,
-			/// <summary>
-			/// Buffer used for drawing, specified many times by application data.
-			/// </summary>
-			/// <remarks>
-			/// <para>
-			/// By using this hint, this BufferObject allocates client memory to set buffer object contents,
-			/// and after done it does not release the client memory, in order to update frequently the
-			/// buffer contents.
-			/// </para>
-			/// </remarks>
-			DynamicCpuDraw = Gl.DYNAMIC_DRAW,
-			/// <summary>
-			/// Buffer used for drawing, specified many times by OpenGL rendering.
-			/// </summary>
-			/// <remarks>
-			/// <para>
-			/// By using this hint, this BufferObject does not allocates client memory.
-			/// </para>
-			/// </remarks>
-			DynamicGpuDraw = Gl.DYNAMIC_COPY,
-			/// <summary>
-			/// Buffer used for reading, specified once by OpenGL rendering.
-			/// </summary>
-			/// <remarks>
-			/// <para>
-			/// By using this hint, this BufferObject allocates client memory to get buffer object contents.
-			/// Client memory will be released at object disposition.
-			/// </para>
-			/// </remarks>
-			StaticRead = Gl.STATIC_READ,
-			/// <summary>
-			/// Buffer used for reading, specified many times by OpenGL rendering.
-			/// </summary>
-			/// <remarks>
-			/// <para>
-			/// By using this hint, this BufferObject allocates client memory to get buffer object contents.
-			/// Client memory will be released at object disposition.
-			/// </para>
-			/// </remarks>
-			DynamicRead = Gl.DYNAMIC_READ
-		}
+		#region Buffer Definition
 
 		/// <summary>
 		/// Get this BufferObject type.
@@ -144,24 +69,43 @@ namespace OpenGL
 		public BufferTargetARB BufferType { get { return (_Type); } }
 
 		/// <summary>
-		/// Get this BufferObject hint.
+		/// BufferObject type, indicatinvg the usage pattern.
 		/// </summary>
-		public Hint BufferHint { get { return (_Hint); } }
+		private readonly BufferTargetARB _Type;
 
 		/// <summary>
-		/// Get this BufferObject total size, in basic machine units.
+		/// Get this BufferObject hint.
+		/// </summary>
+		public BufferObjectHint BufferHint { get { return (_Hint); } }
+
+		/// <summary>
+		/// BufferObject usage hint.
+		/// </summary>
+		private readonly BufferObjectHint _Hint;
+
+		/// <summary>
+		/// Get this BufferObject total size, in bytes.
 		/// </summary>
 		public uint BufferSize
 		{
-			get { return (_Size); }
-			protected set { _Size = value; }
+			get { return (_BufferSize); }
+			protected set { _BufferSize = value; }
 		}
 
 		/// <summary>
-		/// Define this buffer object allocation.
+		/// Size of the buffer object, in basic machine units (bytes).
+		/// </summary>
+		private uint _BufferSize;
+
+		#endregion
+
+		#region Client Memory
+
+		/// <summary>
+		/// Allocate client memory for this BufferObject.
 		/// </summary>
 		/// <param name="size">
-		/// A <see cref="UInt32"/> that determine the size of the buffer object, in bytes.
+		/// A <see cref="UInt32"/> that determine the size of the buffer object client memory, in bytes.
 		/// </param>
 		/// <remarks>
 		/// <para>
@@ -171,13 +115,13 @@ namespace OpenGL
 		/// </remarks>
 		protected void Allocate(uint size)
 		{
+			// Discard previous buffer
 			if (MemoryBuffer != null)
 				MemoryBuffer.Dispose();
-
 			// Allocate memory, if required
-			MemoryBuffer = new AlignedMemoryBuffer(_Size, 16);
+			MemoryBuffer = new AlignedMemoryBuffer(size, DefaultBufferAlignment);
 			// Store buffer object size
-			_Size = size;
+			_BufferSize = size;
 		}
 
 		/// <summary>
@@ -188,58 +132,70 @@ namespace OpenGL
 		/// </param>
 		protected void Reallocate(uint size)
 		{
-			if (MemoryBuffer == null) {
-				Allocate(size);
-				return;
-			} else {
+			if (MemoryBuffer != null) {
 				MemoryBuffer.Realloc(size);
-				_Size = size;
+				_BufferSize = size;
+			} else {
+				Allocate(size);
 			}
 		}
-
-		/// <summary>
-		/// BufferObject type, indicatinvg the usage pattern.
-		/// </summary>
-		private readonly BufferTargetARB _Type;
-
-		/// <summary>
-		/// BufferObject hints.
-		/// </summary>
-		private readonly Hint _Hint;
-
-		/// <summary>
-		/// Size of the buffer object, in basic machine units (bytes).
-		/// </summary>
-		private uint _Size;
 
 		/// <summary>
 		/// The buffer object representation in client memory.
 		/// </summary>
 		internal AlignedMemoryBuffer MemoryBuffer;
-		
+
+		/// <summary>
+		/// The default memory alignment required.
+		/// </summary>
+		protected const uint DefaultBufferAlignment = 16;
+
 		#endregion
 
 		#region Buffer Binding
-		
+
 		/// <summary>
 		/// Bind this BufferObject.
 		/// </summary>
+		/// <param name="ctx">
+		/// The <see cref="GraphicsContext"/> used for binding this BufferObject.
+		/// </param>
+		/// <exception cref="ArgumentNullException">
+		/// Exception thrown if <paramref name="ctx"/> is null.
+		/// </exception>
+		/// <exception cref="ArgumentException">
+		/// Exception thrown if <paramref name="ctx"/> is not current on the calling thread.
+		/// </exception>
+		/// <exception cref="InvalidOperationException">
+		/// Exception thrown if <see cref="GraphicsResource.ObjectName"/> represents an invalid object name.
+		/// </exception>
 		internal virtual void Bind(GraphicsContext ctx)
 		{
+			if (ctx == null)
+				throw new ArgumentNullException("ctx");
+			if (ctx.IsCurrent == false)
+				throw new ArgumentException("not current", "ctx");
 			if (ObjectName == InvalidObjectName)
 				throw new InvalidOperationException("invalid name");
 			
-			// Bind this buffer object
-			Gl.BindBuffer(_Type, ObjectName);
+			if (ctx.Caps.GlExtensions.VertexArrayObject_ARB) {
+				// Bind this buffer object
+				Gl.BindBuffer(_Type, ObjectName);
+			}
 		}
 
 		/// <summary>
 		/// Unbind this BufferObject.
 		/// </summary>
+		/// <param name="ctx">
+		/// The <see cref="GraphicsContext"/> used for unbinding any BufferObject.
+		/// </param>
 		internal virtual void Unbind(GraphicsContext ctx)
 		{
-			// Unbind actually binded buffer object
-			Gl.BindBuffer(_Type, 0);
+			if (ctx.Caps.GlExtensions.VertexArrayObject_ARB) {
+				// Unbind actually binded buffer object
+				Gl.BindBuffer(_Type, InvalidObjectName);
+			}
 		}
 
 		#endregion
@@ -253,17 +209,42 @@ namespace OpenGL
 		/// A <see cref="GraphicsContext"/> required for mapping this BufferObject.
 		/// </param>
 		/// <param name="mask">
-		/// A <see cref="MapMask"/> that specify the map access.
+		/// A <see cref="BufferAccessARB"/> that specify the map access.
 		/// </param>
+		/// <exception cref="ArgumentNullException">
+		/// Exception thrown if <paramref name="ctx"/> is null.
+		/// </exception>
+		/// <exception cref="InvalidOperationException">
+		/// Exception thrown if this BufferObject is already mapped.
+		/// </exception>
+		/// <exception cref="InvalidOperationException">
+		/// Exception thrown if this BufferObject does not exist, and no client memory buffer is allocated.
+		/// </exception>
+		/// <remarks>
+		/// <para>
+		/// This method does not care whether the BufferObject is actually created (i.e. an actual GPU
+		/// buffer is allocated): in the case the GL_ARB_vertex_array_object is not supported, nor
+		/// the buffer is not yet created, the mapped buffer corresponds with the client buffer object memory.
+		/// </para>
+		/// <para>
+		/// After being mapped, the mapped buffer is accessible via <see cref="MapGet{T}(GraphicsContext, long)"/> or
+		/// <see cref="MapSet{T}(GraphicsContext, T, long)"/>, or via <see cref="MappedBuffer"/> as raw pointer.
+		/// </para>
+		/// </remarks>
 		public void Map(GraphicsContext ctx, BufferAccessARB mask)
 		{
+			if (ctx == null)
+				throw new ArgumentNullException("ctx");
+
 			if (IsMapped())
 				throw new InvalidOperationException("already mapped");
 
-			if (Exists(ctx)) {
+			if (Exists(ctx) && ctx.Caps.GlExtensions.VertexArrayObject_ARB) {
 				// Map buffer object data (resident on server)
 				_MappedBuffer = Gl.MapBuffer(_Type, mask);
 			} else {
+				if (MemoryBuffer == null)
+					throw new InvalidOperationException("not buffer defined");
 				// Emulate mapping
 				_MappedBuffer = MemoryBuffer.AlignedBuffer;
 			}
@@ -322,37 +303,50 @@ namespace OpenGL
 
 			return ((T)Marshal.PtrToStructure(new IntPtr(_MappedBuffer.ToInt64()+offset), typeof(T)));
 		}
-		
+
 		/// <summary>
-		/// Unmap this RenderBufferObject data.
+		/// Unmap this BufferObject data.
 		/// </summary>
+		/// <exception cref="ArgumentNullException">
+		/// Exception thrown if <paramref name="ctx"/> is null.
+		/// </exception>
+		/// <exception cref="InvalidOperationException">
+		/// Exception thrown if this BufferObject is not mapped.
+		/// </exception>
+		/// <exception cref="InvalidOperationException">
+		/// Exception thrown if this BufferObject has been corrupted after being unmapped.
+		/// </exception>
 		public void Unmap(GraphicsContext ctx)
-		{	
+		{
+			if (ctx == null)
+				throw new ArgumentNullException("ctx");
 			if (IsMapped() == false)
 				throw new InvalidOperationException("not mapped");
 
-			if (Exists(ctx)) {
+			if (Exists(ctx) && ctx.Caps.GlExtensions.VertexArrayObject_ARB) {
 				// Unmap buffer object data (resident on server)
 				bool uncorrupted = Gl.UnmapBuffer(_Type);
 
 				if (uncorrupted == false)
-					throw new InvalidOperationException("corrupted when unmapping");
+					throw new InvalidOperationException("corrupted buffer");
 			}
 			// Removes reference for mapped buffer data
 			_MappedBuffer = IntPtr.Zero;
 		}
 
 		/// <summary>
-		/// Check whether this RenderBufferObject data is mapped.
+		/// Check whether this BufferObject data is mapped.
 		/// </summary>
-		/// <returns></returns>
+		/// <returns>
+		/// It returns a boolean value indicating whether this BufferObject is mapped.
+		/// </returns>
 		public bool IsMapped()
 		{
 			return (_MappedBuffer != IntPtr.Zero);
 		}
 
 		/// <summary>
-		/// Get the mapped buffer.
+		/// Get the mapped buffer address.
 		/// </summary>
 		protected IntPtr MappedBuffer { get { return (_MappedBuffer); } }
 
@@ -362,19 +356,27 @@ namespace OpenGL
 		private IntPtr _MappedBuffer = IntPtr.Zero;
 
 		#endregion
-		
+
 		#region Get Data
-		
+
+		/// <summary>
+		/// Retrieve the buffer object context
+		/// </summary>
+		/// <param name="ctx"></param>
 		public void GetBufferData(GraphicsContext ctx)
 		{
-			if (MemoryBuffer != null)
-				MemoryBuffer.Dispose();
-			
-			MemoryBuffer = new AlignedMemoryBuffer(_Size, 16);
-			
-			Bind(ctx);
-			
-			Gl.GetBufferSubData(_Type, IntPtr.Zero, _Size, MemoryBuffer.AlignedBuffer);
+			if (ctx == null)
+				throw new ArgumentNullException("ctx");
+			if (ctx.IsCurrent == false)
+				throw new ArgumentException("not current", "ctx");
+
+			if (ctx.Caps.GlExtensions.VertexArrayObject_ARB) {
+				Reallocate(_BufferSize);
+
+				Bind(ctx);
+
+				Gl.GetBufferSubData(_Type, IntPtr.Zero, _BufferSize, MemoryBuffer.AlignedBuffer);
+			}
 		}
 		
 		#endregion
@@ -410,13 +412,21 @@ namespace OpenGL
 		/// <exception cref="ArgumentNullException">
 		/// Exception thrown if <paramref name="ctx"/> is null.
 		/// </exception>
+		/// <exception cref="ArgumentException">
+		/// Exception thrown if <paramref name="ctx"/> is not current on the calling thread.
+		/// </exception>
 		public override bool Exists(GraphicsContext ctx)
 		{
+			if (ctx == null)
+				throw new ArgumentNullException("ctx");
+			if (ctx.IsCurrent == false)
+				throw new ArgumentException("not current");
+
 			// Object name space test (and 'ctx' sanity checks)
 			if (base.Exists(ctx) == false)
 				return (false);
 
-			return (ctx.IsCurrent && Gl.IsBuffer(ObjectName));
+			return (!ctx.Caps.GlExtensions.VertexArrayObject_ARB || Gl.IsBuffer(ObjectName));
 		}
 
 		/// <summary>
@@ -428,8 +438,21 @@ namespace OpenGL
 		/// <returns>
 		/// It returns a valid object name for this BufferObject.
 		/// </returns>
+		/// <exception cref="ArgumentNullException">
+		/// Exception thrown if <paramref name="ctx"/> is null.
+		/// </exception>
+		/// <exception cref="ArgumentException">
+		/// Exception thrown if <paramref name="ctx"/> is not current on the calling thread.
+		/// </exception>
 		protected override uint CreateName(GraphicsContext ctx)
 		{
+			if (ctx == null)
+				throw new ArgumentNullException("ctx");
+			if (ctx.IsCurrent == false)
+				throw new ArgumentException("not current");
+
+			Debug.Assert(ctx.Caps.GlExtensions.VertexArrayObject_ARB);
+
 			return (Gl.GenBuffer());
 		}
 
@@ -440,10 +463,28 @@ namespace OpenGL
 		/// A <see cref="GraphicsContext"/> used for deleting this buffer object name.
 		/// </param>
 		/// <param name="name">
-		/// A <see cref="System.UInt32"/> that specify the object name to delete.
+		/// A <see cref="UInt32"/> that specify the object name to delete.
 		/// </param>
+		/// <exception cref="ArgumentNullException">
+		/// Exception thrown if <paramref name="ctx"/> is null.
+		/// </exception>
+		/// <exception cref="ArgumentException">
+		/// Exception thrown if <paramref name="ctx"/> is not current on the calling thread.
+		/// </exception>
+		/// <exception cref="InvalidOperationException">
+		/// Exception thrown if this BufferObject does not exist.
+		/// </exception>
 		protected override void DeleteName(GraphicsContext ctx, uint name)
 		{
+			if (ctx == null)
+				throw new ArgumentNullException("ctx");
+			if (ctx.IsCurrent == false)
+				throw new ArgumentException("not current");
+			if (Exists(ctx) == false)
+				throw new InvalidOperationException("not existing");
+
+			Debug.Assert(ctx.Caps.GlExtensions.VertexArrayObject_ARB);
+
 			// Delete buffer object
 			Gl.DeleteBuffers(name);
 		}
@@ -454,27 +495,49 @@ namespace OpenGL
 		/// <param name="ctx">
 		/// A <see cref="GraphicsContext"/> used for allocating resources.
 		/// </param>
+		/// <exception cref="ArgumentNullException">
+		/// Exception thrown if <paramref name="ctx"/> is null.
+		/// </exception>
+		/// <exception cref="ArgumentException">
+		/// Exception thrown if <paramref name="ctx"/> is not current on the calling thread.
+		/// </exception>
+		/// <exception cref="InvalidOperationException">
+		/// Exception thrown if this BufferObject has not client memory allocated and the hint is different from
+		/// <see cref="BufferObjectHint.StaticCpuDraw"/> or <see cref="BufferObjectHint.DynamicCpuDraw"/>.
+		/// </exception>
+		/// <exception cref="InvalidOperationException">
+		/// Exception thrown if this BufferObject is currently mapped.
+		/// </exception>
 		protected override void CreateObject(GraphicsContext ctx)
 		{
-			if (_Size == 0)
+			if (ctx == null)
+				throw new ArgumentNullException("ctx");
+			if (ctx.IsCurrent == false)
+				throw new ArgumentException("not current");
+			if (_BufferSize == 0)
 				throw new InvalidOperationException("size not defined");
-			if ((MemoryBuffer == null) && ((_Hint != Hint.StaticCpuDraw) && (_Hint != Hint.DynamicCpuDraw)))
+			if ((MemoryBuffer == null) && ((_Hint != BufferObjectHint.StaticCpuDraw) && (_Hint != BufferObjectHint.DynamicCpuDraw)))
 				throw new InvalidOperationException("contents not defined");
 
 			// Buffer must be bound
 			Bind(ctx);
 
-			// Define buffer object (type, size and hints)
-			Gl.BufferData((int)_Type, _Size, null, (int)_Hint);
+			if (ctx.Caps.GlExtensions.VertexArrayObject_ARB) {
+				if (IsMapped())
+					throw new InvalidOperationException("mapped");
 
-			// Define buffer object contents
-			if (MemoryBuffer != null) {
-				// Provide buffer contents
-				Gl.BufferSubData(_Type, IntPtr.Zero, _Size, MemoryBuffer.AlignedBuffer);
-				// Release memory, if it is not required anymore
-				if (_Hint == Hint.StaticCpuDraw) {
-					MemoryBuffer.Dispose();
-					MemoryBuffer = null;
+				// Define buffer object (type, size and hints)
+				Gl.BufferData((int)_Type, _BufferSize, null, (int)_Hint);
+
+				// Define buffer object contents
+				if (MemoryBuffer != null) {
+					// Provide buffer contents
+					Gl.BufferSubData(_Type, IntPtr.Zero, _BufferSize, MemoryBuffer.AlignedBuffer);
+					// Release memory, if it is not required anymore
+					if (_Hint == BufferObjectHint.StaticCpuDraw) {
+						MemoryBuffer.Dispose();
+						MemoryBuffer = null;
+					}
 				}
 			}
 		}
